🔥 HotSot Week 1-2 Implementation: Critical Production Fixes
## Week 1 — Critical Fixes
- **Idea #4**: Decimal/Money class verified across all services (pricing, billing, compensation, menu)
- **Idea #17**: Pricing service fixed — additive discounts against base, floor check, 50% cap
- **Idea #5**: Compliance engine with @compliance_check/@require_compliance decorators
- **Idea #20**: Billing zero-amount invoice prevention with validate_invoice_amount()
- **Idea #6**: Order payments.py rewritten — RazorpayGateway + Money.to_paise() (no float * 100)

## Week 2 — Architecture Hardening
- **Idea #2/#3**: Contract files created for all 13 missing services
- **Idea #9**: shared/optimistic_locking.py wired into services
- **Idea #16**: JWT revocation via Redis blocklist (fail-closed)
- Compliance decorators on vendor (require_compliance FSSAI+GST) and menu (compliance_check FSSAI)
- Vendor activation endpoint with compliance enforcement

## Key Files Changed
- services/order-service/app/routes/payments.py — Full rewrite with Money + RazorpayGateway
- services/vendor-service/app/routes/vendor.py — @require_compliance on activate
- services/menu-service/app/routes/menu.py — @compliance_check on item creation
- 13 new contracts.py files for contract-first development
- tests/test_comprehensive.py — Full test suite (Money, GST, Pricing, Circuit Breaker, etc.)

+++ services/search-service/app/core/database.py
@@ -21,7 +21,7 @@ Migration Note (GIN Index):
     to_tsvector() inline in queries (slower but functional).
 """
 from sqlalchemy import Column, String, Integer, Float, Boolean, Text, Index
-from sqlalchemy.dialects.postgresql import UUID as PG_UUID, JSONB, TSVECTOR
+from sqlalchemy.dialects.postgresql import UUID as PG_UUID, JSONB, TSVECTOR, Numeric
 from shared.utils.database import BaseModel
 
 
@@ -37,8 +37,8 @@ class SearchIndexModel(BaseModel):
     cuisine = Column(String(50), nullable=True)
     city = Column(String(100), nullable=True)
     is_available = Column(Boolean, default=True)
-    price_min = Column(Float, nullable=True)  # Minimum price in INR
-    price_max = Column(Float, nullable=True)  # Maximum price in INR
+    price_min = Column(Numeric(10, 2), nullable=True)  # Minimum price in INR
+    price_max = Column(Numeric(10, 2), nullable=True)  # Maximum price in INR
     dietary_preference = Column(String(20), nullable=True, index=True)  # VEG / NON_VEG / VEGAN / EGGETARIAN
     search_vector = Column(Text, nullable=True)  # Legacy — kept for backward compat
     search_tsvector = Column(TSVECTOR, nullable=True)  # GIN-indexed full-text vector (set by migration)
